fixing

+++ sources/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: aadyan <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 20:57:42 by aadyan            #+#    #+#             */
-/*   Updated: 2024/08/13 19:06:59 by aadyan           ###   ########.fr       */
+/*   Updated: 2024/08/28 15:07:45 by aadyan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*memset(void *ptr, int value, size_t num)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*p;
 
 	p = ptr;
